requirements.txt
absl-py==1.4.0
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
# backports.zoneinfo==0.2.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
Brotli==1.0.9
cachetools==5.3.1
catalogue==2.0.10
certifi==2022.12.7
cfgv==3.4.0
charset-normalizer==3.0.1
click==8.1.7
cloudpathlib==0.16.0
cmake==3.25.0
confection==0.1.4
contexttimer==0.3.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
datasets==2.14.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.0
diffusers==0.17.1
dill==0.3.7
distlib==0.3.7
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==2.0.1
fairscale==0.4.4
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
gmpy2==2.1.2
google-auth==2.20.0
google-auth-oauthlib==0.4.6
grpcio==1.54.2
hjson==3.1.0
huggingface-hub==0.19.4
identify==2.5.32
idna==3.4
imageio==2.33.0
importlib-metadata==6.6.0
importlib-resources==6.1.1
iniconfig==2.0.0
iopath==0.1.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kaggle==1.5.16
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
lit==15.0.7
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.0
ninja==1.11.1.1
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
peft==0.6.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.0.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.41
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.0.9
pyre-extensions==0.0.23
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.31.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
thinc==8.2.1
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.13.0
transformers==4.35.2
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.4.0
tzlocal==5.2
urllib3==1.26.14
validators==0.22.0
virtualenv==20.21.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
webdataset==0.2.79
webencodings==0.5.1
Werkzeug==2.3.6
xformers==0.0.18
xxhash==3.4.1
yarl==1.9.3
zipp==3.15.0
gradio



# absl-py
# accelerate
# aiohttp
# aiosignal
# altair
# antlr4-python3-runtime
# asttokens
# async-timeout
# attrs
# backcall
# bleach
# blinker
# blis
# braceexpand
# Brotli
# cachetools
# catalogue
# certifi
# cfgv
# charset-normalizer
# click
# cloudpathlib
# cmake
# confection
# contexttimer
# contourpy
# cycler
# cymem
# datasets
# decorator
# decord
# deepspeed
# diffusers
# dill
# distlib
# einops
# et-xmlfile
# exceptiongroup
# executing
# fairscale
# filelock
# fonttools
# frozenlist
# fsspec
# ftfy
# gitdb
# GitPython
# gmpy2
# google-auth
# google-auth-oauthlib
# grpcio
# hjson
# huggingface-hub
# identify
# idna
# imageio
# importlib-metadata
# importlib-resources
# iniconfig
# iopath
# ipython
# jedi
# Jinja2
# joblib
# jsonschema
# jsonschema-specifications
# kaggle
# kiwisolver
# langcodes
# lazy_loader
# lit
# Markdown
# markdown-it-py
# MarkupSafe
# matplotlib
# matplotlib-inline
# mdurl
# mkl-fft
# mkl-random
# mkl-service
# mpmath
# multidict
# multiprocess
# murmurhash
# mypy-extensions
# networkx
# ninja
# nodeenv
# oauthlib
# omegaconf
# opencv-python-headless
# opendatasets
# openpyxl
# packaging
# pandas
# parso
# peft
# pexpect
# pickleshare
# Pillow
# pkgutil_resolve_name
# platformdirs
# pluggy
# portalocker
# pre-commit
# preshed
# prompt-toolkit
# protobuf
# psutil
# ptyprocess
# pure-eval
# py-cpuinfo
# pyarrow
# pyarrow-hotfix
# pyasn1
# pyasn1-modules
# pycocoevalcap
# pycocotools
# pydantic
# pydeck
# Pygments
# pynvml
# pyparsing
# pyre-extensions
# PySocks
# pytest
# python-dateutil
# python-magic
# python-slugify
# pytz
# PyWavelets
# PyYAML
# referencing
# regex
# requests
# requests-oauthlib
# rich
# rpds-py
# rsa
# safetensors
# scikit-image
# scipy
# sentencepiece
# six
# smart-open
# smmap
# spacy
# spacy-legacy
# spacy-loggers
# srsly
# stack-data
# streamlit
# sympy
# tenacity
# tensorboard
# tensorboard-data-server
# tensorboard-plugin-wit
# text-unidecode
# thinc
# tifffile
# timm
# tokenizers
# toml
# tomli
# toolz
# tqdm
# traitlets
# transformers
# triton
# typer
# typing-inspect
# typing_extensions
# tzlocal
# urllib3
# validators
# virtualenv
# wasabi
# watchdog
# wcwidth
# weasel
# webdataset
# webencodings
# Werkzeug
# xformers
# xxhash
# yarl
# zipp
